requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
watchdog==3.0.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dateutil==2.8.2
netaddr==0.8.0
psutil==5.9.6
colorama==0.4.6
requests==2.31.0
pyyaml==6.0.1
schedule==1.2.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
flask==2.3.3
werkzeug==2.3.7
paramiko==3.3.1
ipaddress==1.0.23
rdpy==0.1.0
impacket==0.11.0
# Additional dependencies for centralized configuration
ipaddress==1.0.23
hashlib
threading
socket
json
datetime
pathlib
typing
dataclasses
concurrent.futures 